approval_notification/requirements.txt
google-cloud-secret-manager==2.0.0
google-cloud-logging==1.15.1
google-cloud-firestore==1.9.0
slackclient==2.9.0
pytest==6.1.1
